+++ apps/autotest/models/TST/_initialData/020_fillData.js
@@ -13,7 +13,9 @@ module.exports = function (session) {
   csvLoader.loadSimpleCSVData(conn, __dirname + '/ubm_enum-TST.csv', 'ubm_enum', 'eGroup;code;name;sortOrder'.split(';'), [0, 1, 2, 3], 1)
 
   console.info('\tFill TST model dictionary')
-  csvLoader.loadSimpleCSVData(conn, __dirname + '/tst_dictionary-TST.csv', 'tst_dictionary', 'ID;code;caption;filterValue;booleanColumn;currencyValue;floatValue'.split(';'), [0, 1, 2, 3, 4, 5, 6], 1)
+  csvLoader.loadSimpleCSVData(conn, __dirname + '/tst_dictionary-TST.csv', 'tst_dictionary',
+    'ID;code;caption;filterValue;booleanColumn;currencyValue;floatValue;jsonColumn'.split(';'),
+    [0, 1, 2, 3, 4, 5, 6, (arr) => arr[7] === '' ? null : arr[7]], 1)
 
   console.info('\tFill TST main data')
   csvLoader.loadSimpleCSVData(conn, __dirname + '/tst_maindata-TST.csv', 'tst_maindata', 'code;caption;nonNullDict_ID;nullDict_ID;enumValue;booleanValue;manyValue;dateTimeValue'.split(';'), [0, 1, 2, 3, 4, 5, 6, 7], 1)