+++ packages/adminui-pub/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [5.6.0]
+### Added
+ - attributes of type `Json` now supported by adminUI
+ - New @cfg `valueIsJson` added to `UBCodeMirror`. If true value can be plain JS object. In other case - string 
+
 ## [5.5.8]
 ### Fixed
  - new version creation for *cached* entities with `dataHistory` mixin