+++ packages/cs-shared/CHANGELOG.md
@@ -4,4 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [5.0.0]
+## [5.0.10]
+### Added
+ - new convert rule is added for attributes of type `Json` in `getConvertRules` function  
+
+## [5.0.9]
+### Added 
+ - `UBEntity.getEntityAttributeInfo` in case of request to inner keys of Json type attribute
+ will return actual Json attribute in `parentAttribute` and `attribute: ubndefined`
+ 
+### Changed
+ - `UBEntity.getEntityAttributeInfo` speed up from x10 to x100 (avoid calling String.split if not necessary)
+ - `UBEntity.getEntityAttributeInfo` will return additional parameter `parentAttribute`
+
+### Fixed 
+ - `UBEntity.getEntityAttributeInfo` will return correct entity (listed after @) for
+ cases `domain.get('org_unit').getEntityAttributeInfo('parentID.code@org_department')`.
+ Previous implementation return `org_unit` for such query. 
+